USB: add workaround for some rom driver bugs
- usb_rom_composite: fixed
- usb_rom_vcom: fixed
- usb_rom_msc: fixed

Without this fix, enumeration is very flaky (especially on linux?).
With these workarounds, enumeration seems to be very stable and without errors.
Some more testing on different platforms may be required though..

The fixes are based on example code and hints from multiple errata sheets
for the LPC11Uxx and similar platforms.